🔧 Refactor response schema handling & simplify validation in tests
Refactor the handling of response schemas in OpenAPI by removing
`RestApiOperationResponseSchema` class and directly using
`JsonDocument` for storing schemas. Update related methods and
properties accordingly. Additionally, simplify schema validation in
PluginTests by removing redundant checks and replacing them with a
single call to the IsValid() method, making the test code more concise
and easier to maintain.
